fix: contrast计算显式Float

+++ app/src/main/java/com/wxjxpp/neiro/feature/player/PlayerDetailScreen.kt
@@ -333,8 +333,8 @@ val palette = bmp.extractVividPalette()
             val argb = col.toArgb()
             val hsv = FloatArray(3).also { android.graphics.Color.colorToHSV(argb, it) }
             if (hsv[1] < 0.18f) return@mapNotNull null
-            val contrast = kotlin.math.abs(lum(argb) - canvasLum)
-            Cand(hsv, contrast * (0.35f + hsv[1]).toFloat())
+            val contrast = kotlin.math.abs((lum(argb) - canvasLum).toDouble()).toFloat()
+            Cand(hsv, (contrast * (0.35f + hsv[1])).toFloat())
         }.maxByOrNull { it.score }
         if (best == null) immersiveScheme.primary else {
             val hsv = floatArrayOf(best.hsv[0], best.hsv[1], best.hsv[2])
